+++ app/src/main/java/com/link_up/matching_manager/ui/my/bookmark/MyBookmarkMatchFragment.kt
@@ -14,8 +14,8 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.link_up.matching_manager.databinding.MyBookmarkMatchFragmentBinding
 import com.link_up.matching_manager.ui.match.MatchDataModel
 import com.link_up.matching_manager.ui.match.MatchDetailActivity.Companion.detailIntent
-import com.link_up.matching_manager.ui.my.my.MyViewModelFactory
-import com.link_up.matching_manager.ui.my.my.MyViewModel
+import com.link_up.matching_manager.ui.my.my.MyPostViewModelFactory
+import com.link_up.matching_manager.ui.my.my.MyPostViewModel
 import com.google.gson.Gson
 
 class MyBookmarkMatchFragment : Fragment() {
@@ -23,8 +23,8 @@ class MyBookmarkMatchFragment : Fragment() {
     private var _binding: MyBookmarkMatchFragmentBinding? = null
     private val binding get() = _binding!!
 
-    private val viewModel: MyViewModel by viewModels {
-        MyViewModelFactory()
+    private val viewModel: MyPostViewModel by viewModels {
+        MyPostViewModelFactory()
     }
 
     private val resultLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
