+++ app/src/main/java/com/link_up/matching_manager/ui/my/team/MyTeamRecruitEditActivity.kt
@@ -18,9 +18,9 @@ import coil.load
 import com.link_up.matching_manager.R
 import com.link_up.matching_manager.databinding.MyTeamRecruitEditActivityBinding
 import com.link_up.matching_manager.ui.my.my.MyEvent
-import com.link_up.matching_manager.ui.my.my.MyViewModel
+import com.link_up.matching_manager.ui.my.my.MyPostViewModel
 import com.link_up.matching_manager.ui.my.match.MyMatchMenuBottomSheet
-import com.link_up.matching_manager.ui.my.my.MyViewModelFactory
+import com.link_up.matching_manager.ui.my.my.MyPostViewModelFactory
 import com.link_up.matching_manager.ui.team.TeamItem
 import com.link_up.matching_manager.ui.team.TeamWritingActivity
 import com.link_up.matching_manager.ui.team.bottomsheet.TeamCalenderBottomSheet
@@ -36,8 +36,8 @@ class MyTeamRecruitEditActivity : AppCompatActivity() {
 
     private val sharedViewModel: TeamSharedViewModel by viewModels()
 
-    private val viewModel: MyViewModel by viewModels {
-        MyViewModelFactory()
+    private val viewModel: MyPostViewModel by viewModels {
+        MyPostViewModelFactory()
     }
 
     private val reference: StorageReference = FirebaseStorage.getInstance().reference
